% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{constructionError}
\alias{constructionError}
\title{Calculate Construction Error of 2D Plot (tSNE/UMAP/PCA)
Used by \code{getHighestCorrelation} function}
\usage{
constructionError(original_data, est_dist)
}
\description{
Calculate Construction Error of 2D Plot (tSNE/UMAP/PCA)
Used by \code{getHighestCorrelation} function
}
