% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dim_Red}
\alias{dim_Red}
\title{Dimension reduction}
\usage{
dim_Red(
  sce,
  annot = FALSE,
  uncert = 0.2,
  perplex = 30,
  num_threads = 1,
  reduction = "umap",
  n.neighbors = 20,
  min.dist = 0.5,
  spread = 1,
  verbose = FALSE,
  seed = 42,
  a = NULL,
  b = NULL,
  resolution = 0.5,
  npcs = 50
)
}
\arguments{
\item{sce}{sce SingleCellExperiment object returned by \code{preprocess_Data} function}

\item{annot}{Annotation (eg. cell type) of each single cell if available, default is FALSE}

\item{uncert}{Threshold of uncertainty in modelling / class assignment. Beyong the set threhold, cells will be marked as uncertain. Recommended value is 0.5}

\item{perplex}{Perpexity parameter for tSNE Set it to -1 if you want the value to be calculated according to the formula described in paper. Recommended value is -1}

\item{reduction}{Value can be 'umap' or 'tsne' or 'pca'. Set value according to the dimension reduction techniqe you want to use.}

\item{n.neighbors}{Parameter to create neighborhood graph for UMAP/tSNE}

\item{min.dist}{min.dist Parameter for UMAP}

\item{spread}{Parameter for UMAP. Recommended value is 1.5}

\item{a}{Parameter for UMAP}

\item{b}{Parameter for UMAP}

\item{resolution}{Resolution value to create clusters for Leiden clustering.}

\item{npcs}{Number of Principal Components to be used for UMAP/tSNE}
}
\value{
Returns the Embeddings list containing UMAP/tSNE/PCA derived latent dimensions plus class labels obtained using GMM
plus marking of cells as certain or uncertain plus supplied annotations along with the probability values
}
\description{
Performs dimension reduction using UMAP/tSNE/PCA. Then fits GMM on the clusters in latent dimensions.
}
\examples{
DR_list = dim_Red(sce  = sce ,annot = annot, num_threads = 1,resolution = 0.5,
reduction = 'umap',verbose = TRUE, perplex=-1,
n.neighbors = 30, uncert = 0.5, spread = 1.5)
}
