% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{do_umap_seurat}
\alias{do_umap_seurat}
\title{Perform UMAP. Used by \code{dim_Red} function.}
\usage{
do_umap_seurat(
  sce,
  annot = FALSE,
  n.neighbors = 20,
  num_threads = 1,
  scale = FALSE,
  seed = 42,
  min.dist = 0.5,
  spread = 1,
  a = NULL,
  b = NULL,
  resolution = 0.5,
  npcs = 50
)
}
\description{
Perform UMAP. Used by \code{dim_Red} function.
}
