% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_cells_genes}
\alias{filter_cells_genes}
\title{Filter out poor quality cells and genes}
\usage{
filter_cells_genes(
  raw_Data,
  min_Reads = 1,
  min_Cell = 0.1,
  min_Gene = 1000,
  log = T
)
}
\description{
Filter out poor quality cells and genes using given thresholds
}
