% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getHighestCorrelation}
\alias{getHighestCorrelation}
\title{Finds correlation between original projection and latent dimensions}
\usage{
getHighestCorrelation(
  sce,
  reduced,
  cl_res,
  n_gene_seq = seq(10, 1000, by = 10),
  sample_cells = TRUE
)
}
\arguments{
\item{sce}{sce SingleCellExperiment object returned by \code{preprocess_Data} function}

\item{reduced}{List object of latent dimensions returned by \code{dim_Red} function}

\item{cl_res}{List object returned by \code{rele_Cells} function}

\item{n_gene_seq}{Sequence of number of genes from 10 to desired number (default 1000), by a periodic gap of 10}

\item{sample_cells}{Boolean value parameter. TRUE indicates that sampled cells (by GMM and random selection) will be used to calculate correlation. FALSE indicates all cells will be used. Recommended value is TRUE.}
}
\value{
Returns the list with correlation value for the genes. Considering 'n' genes with the highest correlation value will give us the set of candidate genes that explain the latent projection in the best way
}
\description{
Calculates Spearman correlation coefficient of genes in latent dimension with the original projection based on euclidean distance.
}
