% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getRelevantGenes}
\alias{getRelevantGenes}
\title{Finds top N relevant genes of the latent dimension}
\usage{
getRelevantGenes(
  cl_res,
  nTop = 500,
  ori = c(),
  n_gene_seq = seq(10, 1000, by = 10),
  genes
)
}
\arguments{
\item{cl_res}{List object returned by \code{rele_Cells} function}

\item{nTop}{Number of top relevant genes to be considered. Set it to 0 if you want to find it using using maximum correlation. Recommended value is 500.}

\item{ori}{List with correlation values returned by \code{getHighestCorrelation} function}

\item{n_gene_seq}{Sequence of number of genes from 10 to desired number (default 1000), by a periodic gap of 10}
}
\value{
Returns list with top relevant genes
}
\description{
Uses \code{getHighestCorrelation} to find top N genes influencing the UMAP/tSNE
}
