% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{preprocess_Data}
\alias{preprocess_Data}
\title{Preprocess the raw data}
\usage{
preprocess_Data(
  raw_data,
  gene_list,
  annot,
  nfeatures = "all",
  min_Reads = 5,
  min_Cell = 0.05,
  min_Gene = 1000
)
}
\arguments{
\item{raw_data}{Count or FPKM data for single cell transcriptomes (genes should be on rows)}

\item{gene_list}{the gene names for the dataset}

\item{annot}{the cell labels/annotations/ground truth. If it is unknown, then consecutive numbers starting from 1 wil be used as cell names.}

\item{nfeatures}{The number of variable features that Seurat will use to scale the normalized the data. It can be a number or "all", in which case, all genes after filtering will be retained. We recommend keeping the value as "all".}

\item{min_Reads}{A gene is considered detected if it has at least \code{min_Reads} number of reads in certain percentage of all cells}

\item{min_Cell}{Minimum \code{min_Cells} fraction of all cells should have at least \code{min_Reads} number of reads to be considered as reliably expressed}

\item{min_Gene}{A cell should have at least \code{min_Gene} number of expressed genes}
}
\value{
SingleCellExperiment object with the bad cells removed
}
\description{
Filters out poor quality cells and genes by calling the \code{filter_cells_genes} function.Then normalizes and scales the data.
}
\examples{
library(SingleCellExperiment)
counts <- matrix(rpois(500, lambda = 2), ncol=1000, nrow=2000)
genes <- 1:nrow(counts)
annot <- 1:ncol(counts)
sce <- preprocess_Data(counts, gene_List = genes, annot = annot, nfeatures = "all",  min_Cell = 0.1, min_Reads=0.001, min_Gene = 100)
}
