% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rele_Cells}
\alias{rele_Cells}
\title{Subsamples relevant cells from each cluster}
\usage{
rele_Cells(
  dim_Red_List,
  sce,
  min_Cell = 20,
  n_Trees = 1000,
  num_threads = 1,
  choose_random_cells = TRUE,
  min_randomCell_perCluster = 20,
  seed = 42
)
}
\arguments{
\item{dim_Red_List}{The list returned by \code{dim_Red} function}

\item{sce}{SingleCellExperiment object returned by \code{preprocess_Data} function}

\item{min_Cell}{No of cells to be sampled from each unsupervised group found by GMM modelling}

\item{n_Trees}{No of trees to be populated by the Random Forest algorithm}

\item{num_threads}{Number of threads to run Random Forest}

\item{choose_random_cells}{Can be TRUE or FALSE. Set it to true if you need cells to sampled randomly in addition to the cells selected by GMM Recommended value is TRUE.}

\item{min_randomCell_perCluster}{Number of cells to be selected randomly from each cluster}

\item{seed}{seed value to be set for random sampling and for Random Forest classifier. Default value is 42}
}
\value{
A list containing both gene indices (var_ind), ordered by importance score, and associated important scores (importance). Aslo returns list of cells selected from each cluster (cells_ind).
}
\description{
Feature importance is learned from subsampled cells of each cluster. Finds relevant genes(features) from 2D projections by posing it as a classification problem.
}
\examples{
rele_Cells = rele_Cells(dim_Red_List  = DR_list ,
sce = sce, min_Cell=20,
n_Trees = 1000,num_threads = 4,  
choose_random_cells = TRUE, 
min_randomCell_perCluster = 20, seed = 42)
}
