% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vis_2D_truth}
\alias{vis_2D_truth}
\title{UMAP/tSNE plot to be visualized with cell annotations}
\usage{
vis_2D_truth(dim_Red_List, x = "UMAP 1", y = "UMAP 2")
}
\arguments{
\item{dim_Red_List}{The list returned by \code{dim_Red} function}

\item{x}{The label for X axis. Can be UMAP 1, tSNE 1 or PCA 1 (or any other label according to o user's choice)}

\item{y}{The label for Y axis. Can be UMAP 2, tSNE 2 or PCA 2 (or any other label according to o user's choice)}
}
\value{
plot
}
\description{
UMAP/tSNE plot to be visualized with cell annotations
}
